First pass of changing Select and TextField to typescript
