+++ packages/design-system/src/components/TextField/TextInput.test.tsx
@@ -1,8 +1,8 @@
 import { mount, shallow } from 'enzyme';
 import React from 'react';
-import TextInput from './TextInput';
+import TextInput, { TextInputProps, OmitProps } from './TextInput';
 
-const defaultProps = {
+const defaultProps: Omit<React.ComponentPropsWithRef<'textarea'>, OmitProps> & Omit<React.ComponentPropsWithRef<'input'>, OmitProps> & TextInputProps = {
   name: 'spec-field',
   setRef: jest.fn(),
   id: '1',
