+++ packages/design-system/src/components/TextField/maskHelpers.ts
